requirements.txt
# ═══ FILE: requirements.txt ═══
# Production Dependencies for Aira Voice Call

fastapi==0.110.0
uvicorn[standard]==0.27.1
python-multipart==0.0.9
python-dotenv==1.0.1
httpx==0.27.0
edge-tts==6.1.10
psutil==5.9.8
pydub==0.25.1
gTTS==2.5.1
jinja2==3.1.3
cartesia==1.3.1
# Groq & OpenRouter use httpx for lowest latency
